feat: ✨ carga archivos planos (#35)
* feat ✨ Add CSV seeding functionality for Country entity and include Countries.csv data

* feat(migrations): Update Establishments table constraints and nullability

- Added migration to fix check constraints for DisaCode, MicroRedCode, RedCode, SectorCode, and UniqueCode to allow negative values.
- Updated the nullability of DisaCode, MicroRedCode, and RedCode to be nullable in the Establishments table.
- Adjusted the model snapshot to reflect the changes in the Establishments entity.

* feat ✨ Add Ethnicity mapping and CSV seeding for ethnicities

* feat ✨ Add ExtraCondition mapping and CSV seeding for ethnicities

* feat ✨ Add Financier mapping and CSV seeding for ethnicities

* feat ✨ Add HealthServiceUnit mapping and CSV seeding

* feat ✨ Add DocumentType entity configuration, migration, and model snapshot

* feat ✨ Add DocumentType mapping and CSV seeding

* fix :bug: Update MedicalProcedure entity: change Code type to double and increase Description max length to 255

* feat ✨ Add MedicalProcedure mapping and CSV seeding

* fix: :bug: Update ProfessionalCouncil entity: modify Description max length and adjust Code check constraint

* feat ✨ Add MedicalProcedure mapping and CSV seeding

* feat ✨ Add Profession mapping and CSV seeding

* feat ✨ Add SisProcedure mapping and CSV seeding

* feat ✨ Add Tariff mapping and CSV seeding

+++ src/DE.Domain/Entities/MedicalProcedure.cs
@@ -3,6 +3,6 @@ namespace DE.Domain.Entities;
 public class MedicalProcedure
 {
     public int Id { get; set; }
-    public int Code { get; set; }
+    public double Code { get; set; }
     public string Description { get; set; } = default!;
 }
